+++ src/test/api/ant-manager-test.ts
@@ -4,7 +4,7 @@ import { ApiModelConfig } from '../../api/config/api-model-config';
 import { AntModel } from '../../model/ant-model';
 import { ITest } from '../../testapi/api/ITest';
 import { RedisWrapper } from '../primary/RedisWrapper';
-import { MinimalAntManager } from './MinimalAntManager';
+import { MinimalAntManager } from './minimal-ant-manager';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
