+++ src/test/api/ant-model-manager-test.ts
@@ -4,7 +4,7 @@ import { Entity } from '../../model/entity';
 import { PrimaryModelManager } from '../../persistence/primary/primary-model-manager';
 import { ITest } from '../../testapi/api/ITest';
 import { RedisWrapper } from '../primary/RedisWrapper';
-import { MinimalAntModelManager } from './MinimalAntModelManager';
+import { MinimalAntModelManager } from './minimal-ant-model-manager';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
