+++ src/test/primary/model-manager-test.ts
@@ -10,8 +10,8 @@ import { SingleResultPrimaryQueryManager } from '../../persistence/primary/query
 import { AntJsModelManagerGenerator } from '../../testapi/api/generator/AntJsModelManagerGenerator';
 import { ITest } from '../../testapi/api/ITest';
 import { SecondaryEntityManagerMock } from '../../testapi/api/secondary/SecondaryEntityManagerMock';
-import { MultipleResultQueryByFieldManager } from './query/MultipleResultQueryByFieldManager';
-import { SingleResultQueryByFieldManager } from './query/SingleResultQueryByFieldManager';
+import { MultipleResultQueryByFieldManager } from './query/multiple-result-query-by-field-manager';
+import { SingleResultQueryByFieldManager } from './query/single-result-query-by-field-manager';
 import { RedisWrapper } from './RedisWrapper';
 
 const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
